feat: CreatProduct Ok

+++ Controllers/Product.cs
@@ -6,7 +6,7 @@ namespace Controller
 {
     public class Product
     {
-        public static void AddProduct(string productName, int productValue)
+        public static void AddProduct(string productName, double productValue)
         {
             if (productName == null || productValue == 0)
             {
@@ -17,7 +17,11 @@ namespace Controller
             new Model.Product(productName, productValue);
         }
 
-        public static void UpdateProduct(int productId, string productName, int productValue)
+        public static void UpdateProduct(
+            int productId,
+            string productName,
+            double productValue
+        )
         {
             Model.Product.UpdateProduct(productId, productName, productValue);
         }
